+++ core/terraform/variables.tf
@@ -174,6 +174,12 @@ variable "enable_airlock_malware_scanning" {
   description = "If False, Airlock requests will skip the malware scanning stage"
 }
 
+variable "enable_airlock_email_check" {
+  type        = bool
+  default     = false
+  description = "If True, prior to airlock requests creation will check users have email addresses"
+}
+
 variable "firewall_sku" {
   description = "Azure Firewall SKU"
   type        = string